pyproject.toml
[tool.poetry]
name = "jet_substructure"
version = "0.1.0"
description = "Jet substructure analysis code"
authors = ["Raymond Ehlers <raymond.ehlers@cern.ch>"]
license = "BSD-3-Clause"
include = [
    "jet_substructure/py.typed",
]

[tool.poetry.scripts]
embedPythia = "jet_substructure.analysis.analyze_tree:embed_pythia_entry_point"
splitTree = "jet_substructure.base.helpers:split_tree_entry_point"
mergeROOTFiles = "jet_substructure.base.helpers:merge_ROOT_files_entry_point"
downloadTrainWithConfig = "jet_substructure.base.download_train_with_config:entry_point"
skimDataset = "jet_substructure.analysis.skim_to_flat_tree:skim_entry_point"
processEmbeddingSkim = "jet_substructure.analysis.data_frame:process_embedding_skim_entry_point"
processEmbeddingSkimCpp = "jet_substructure.cpp.data_frame:embed_pythia_entry_point"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.1"
uproot3 = "^3"
uproot = "^4"
pandas = "^1.0.1"
attrs = "^20.3.0"
matplotlib = "^3.1.3"
boost_histogram = "^0.11"
pachyderm = {path = "external/pachyderm", develop = true}
h5py = "^3.1"
coloredlogs = "^14.0"
dask = {extras = ["dataframe"], version = "^2.14.0"}
enlighten = "^1.4.0"
pathos = "^0.2.5"
seaborn = "^0.11"
awkward0 = "^0"
awkward = "^1"
pyarrow = "^2.0"
numba = "^0.51.0"
parsl = {extras = ["monitoring"], version = "^1.0.0"}
mplhep = "^0.2"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
pytest = "^6.1"
flake8 = "^3.7.9"
flake8-colors = "^0.1.6"
black = "^20.8b1"
pre-commit = "^2.0.1"
yamllint = "^1.20.0"
isort = {extras = ["toml"], version = "^5.6"}
IPython = "^7.19"
jupyterlab = "^2.2.7"
jupyter_nbextensions_configurator = "^0.4.1"
# Not explictily enabled because it needs graphviz to be install separately via apt, brew, etc.
#pygraphviz = "^1.5"
networkx = "^2.4"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2
verbose = true
skip = ["docs/conf.py"]
skip_glob = ".*"

[tool.black]
line-length = 120
target-version = ["py37"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
